% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_cells_index}
\alias{get_cells_index}
\title{Getting the column index of interesting cells}
\usage{
get_cells_index(h5f, variants, int_ngt, cell_index = NULL)
}
\arguments{
\item{h5f}{h5f}

\item{variants}{A named vector.
You should get the value from the \code{id} column after calling \code{get_variants}.
If you don't provide the names of the variants, R will do a bad job for you.}

\item{int_ngt}{A numeric vector. Specifies genotypes of interest.}

\item{cell_index}{A numeric vector. Use this if you want to restrict the search to a subset of cells.}
}
\value{
Either a numeric vector or a named list of numeric vectors.
}
\description{
get_cells_index finds which cells are called with the queried NGT for all given variants.
}
\examples{
#Retrieving the NGT matrix for IDH1_R132H and NPM1c
\dontrun{

#Want to find which cells are mutated for IDH1_R132H and NPM1c
interesting_variants <- c(IDH1_R132H = "chr2:209113112:C/T",
                          NPM1c = "chr5:170837543:C/CTCTG")
get_cells_index(h5f,
                variants = interesting_variants,
                int_ngt = c(1, 2))

#Want to find which cells are unknown for IDH1_R132H
get_cells_index(h5f,
                variants = c(IDH1_R132H = "chr2:209113112:C/T"),
                int_ngt = 3)

}
}
