% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_flt3itd}
\alias{get_flt3itd}
\title{Getting FLT3 internal tandem duplication (ITD) variants}
\usage{
get_flt3itd(
  h5f,
  format = c("data.table", "GRanges"),
  gr_coordinate_within = GRanges(seqnames = "chr13", ranges = IRanges(start = 28607990,
    end = 28608531)),
  insertion_size = c(6, 180)
)
}
\arguments{
\item{h5f}{h5f}

\item{format}{\code{data.table} by default.}

\item{gr_coordinate_within}{A \code{GRanges} object.
Specifies the genomic coordinates to search for ITD mutations in the FLT3 gene.
Currently uses hg19 coordinates as reference.}

\item{insertion_size}{A numeric vector with exactly two values.
Specifies the minimum and maximum insertion size (in basepairs) introduced by an ITD mutation.}
}
\value{
A \code{data.table} or a \code{GRanges} object with FLT3-ITD variants.
}
\description{
get_flt3itd retrieves FLT3 ITD variants in an .h5 file
}
