% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_protein_ids}
\alias{get_protein_ids}
\title{Get protein ids}
\usage{
get_protein_ids(h5f)
}
\arguments{
\item{h5f}{h5f}
}
\value{
String
}
\description{
Retrieves the protein ids in a h5 file
}
\examples{
ids <- get_protein_ids(h5f)
}
