% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_variants}
\alias{get_variants}
\title{Getting variants called by haplotype caller in tapestri pipeline}
\usage{
get_variants(h5f, format = c("data.table", "GRanges"))
}
\arguments{
\item{h5f}{h5f}

\item{format}{\code{data.table} by default.}
}
\description{
get_variant retrieves all the dna variants in an .h5 file as a data.table or GRanges
}
\examples{
path <- "path/to/.h5"
h5f <- read_h5f(path)
feature_data <- get_variants(h5f)
}
