% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_variants_index}
\alias{get_variants_index}
\title{Getting the row index of interesting variants}
\usage{
get_variants_index(h5f, variants, sort = TRUE)
}
\arguments{
\item{h5f}{h5f}

\item{variants}{A named vector.
You should get the value from the \code{id} column after calling \code{get_variants}.
If you don't provide the names of the variants, R will do a bad job for you.}

\item{sort}{TRUE or FALSE. Whether to sort indexes by genomic coordinates. Defaults to TRUE.}
}
\value{
A named numeric vector
}
\description{
Given a vector of named variants, get_variants_index returns their row indexes in /assays/dna_variants/layers of an h5f.
It is useful to index which variants (rows) to return when retrieving NGT matrices.
}
\examples{
#Retrieving the NGT matrix for IDH1_R132H and NPM1c
\dontrun{
interesting_variants <- c(IDH1_R132H = "chr2:209113112:C/T", NPM1c = "chr5:170837543:C/CTCTG")
index <- get_variants_index(h5f, interesting_variants)
read_dna_variants(h5f, index_variants = index, included_assays = "NGT")
}
}
