% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_protein_by_clones}
\alias{plot_protein_by_clones}
\title{Plot protein abundance across clones}
\usage{
plot_protein_by_clones(
  h5f,
  index_cells_list,
  select_proteins = NULL,
  geom = c("ridges", "boxplot")
)
}
\arguments{
\item{h5f}{h5f}

\item{index_cells_list}{A named list of integers.
The names are the names of the clones.
The integers are the index of cells that belong to the corresponding clones.}

\item{select_proteins}{A character vector. Specifies the proteins to plot.
To see available options, run \code{rhdf5::h5read(h5f, "/assays/protein_read_counts/ca/id")}}

\item{geom}{Either \code{ridges} or \code{boxplot}. Default is \code{ridges}.}
}
\value{
A named list of ggplot objects
}
\description{
Uses ggplot framework to plot protein abundance across groups of cells (aka, clones).
}
\examples{
#Compare CD34 expression between two clones.
#clone1: IDH1_R132H single mutant
#clone2: IDH1_R132H & NPM1c double mutant


\dontrun{
idh1_het <- get_cells_index(h5f, variants = c(IDH1_R132H = "chr2:209113112:C/T"), int_ngt = "1")
npm1c <- get_cells_index(h5f, variants = c(NPM1c = "chr5:170837543:C/CTCTG"), int_ngt = "1")
npm1_wt <- get_cells_index(h5f, variants = c(NPM1c = "chr5:170837543:C/CTCTG"), int_ngt = "0")

index_clone1 <- intersect(idh1_het, npm1_wt)
index_clone2 <- intersect(idh1_het, npm1c)
index_cells_list <- list(clone1 = index_clone1, clone2 = index_clone2)

plot_protein_by_clones(h5f, index_cells_list, select_proteins = "CD34", geom = "ridges")
}
}
