% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_upset_with_variants}
\alias{plot_upset_with_variants}
\title{Plot an upsetplot}
\usage{
plot_upset_with_variants(
  h5f,
  variants,
  unknown = c("unknown_to_ref", "unknown_discard"),
  return_data = FALSE,
  index_cell = NULL,
  nsets = 5
)
}
\arguments{
\item{h5f}{h5f}

\item{variants}{A named vector.
You should get the value from the \code{id} column after calling \code{get_variants}.
If you don't provide the names of the variants, R will do a bad job for you.}

\item{unknown}{A String. Specifies if NGT = 3 should be considered as reference or discarded for plotting.}

\item{return_data}{Boolean. Defaults to FALSE. If TRUE, returns the NGT data.table with cells as rows and variants as columns.}

\item{index_cell}{A numeric vector. Use this if you want to restrict the search to a subset of cells.}

\item{nsets}{An integer. How many sets (ie, variants) of the supplied variants to plot.}
}
\description{
Upset plot to visualize which variants co-occur.
}
\examples{
path <- "path/to/.h5"
h5f <- read_h5f(path)

interesting_variants <- c(IDH1_R132H = "chr2:209113112:C/T", NPM1c = "chr5:170837543:C/CTCTG")
plot_upset_with_variants(h5f, variants = interesting_variants)
}
