% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_vaf_by_clones}
\alias{plot_vaf_by_clones}
\title{Plot VAF of variants across clones}
\usage{
plot_vaf_by_clones(h5f, index_cells_list, index_variants, return_data = FALSE)
}
\arguments{
\item{h5f}{h5f}

\item{index_cells_list}{A named list of integers.
The names are the names of the clones.
The integers are the index of cells that belong to the corresponding clones.}

\item{return_data}{Boolean. If \code{TRUE}, returns the tidy-data used for plotting}

\item{indx_variants}{See \link[tapestri.tools]{get_variants_index}}
}
\value{
A named list of ggplot objects, or the underlying tidy-data
}
\description{
An in-house function used to identify multiplets in sequencing.
}
\examples{
#Compare IDH1_R132H VAF between wild type and mutant cells

\dontrun{
idh1mt <- get_cells_index(h5f, variants = c(IDH1_R132H = "chr2:209113112:C/T"), int_ngt = 1)
idh1wt <- get_cells_index(h5f, variants = c(IDH1_R132H = "chr2:209113112:C/T"), int_ngt = 0)
index_cells_list <- list(idh1wt = idh1wt, idh1mt = idh1mt)

plot_vaf_by_clones(h5f,
  index_cells_list = indx_ls,
  index_variants = get_variants_index(h5f, c(IDH1_R132H = "chr2:209113112:C/T"))
  )
}
}
