% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_amplicons_bed}
\alias{read_amplicons_bed}
\title{Read tapestri amplicons}
\usage{
read_amplicons_bed(bed, format = c("GRanges", "data.table"))
}
\arguments{
\item{bed}{path to a 4 column .bed file.
The four columns are: chrom, chromStart, chromEnd, name.}

\item{format}{Either \code{GRanges} or \code{data.table}. Defaults to \code{GRanges}.}
}
\description{
Read bed file describing coverage of tapestri amplicons
}
\examples{
\dontrun{
  gr_amplicons <- read_amplicons_bed("path/to/bed")
}
}
