% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_assays_variants}
\alias{read_assays_variants}
\title{Reading the assays of variants}
\usage{
read_assays_variants(
  h5f,
  included_assays = c("AF", "DP", "FILTER_MASK", "GQ", "NGT", "RGQ"),
  index_variants = NULL,
  index_cells = NULL,
  format = c("list", "RangedSummarizedExperiment")
)
}
\arguments{
\item{h5f}{h5f}

\item{included_assays}{A character vector. Selects which assays to retrieve.
Possible values are \describe{
  \item{AF}{variant allele frequencies (VAF) per cell and per variant}
  \item{DP}{read counts per cell and per variant}
  \item{FILTER_MASK}{binary value}
  \item{GQ}{genotype quality values per cell and per variant}
  \item{NGT}{genotypes per cell and per variant \itemize{
    \item 0 = Reference (Wildtype)
    \item 1 = Heterozygous Mutant
    \item 2 = Homozygous Mutant
    \item 3 = Missing Data
    }
  }
  \item{RGQ}{???}
}}

\item{index_variants}{A named numeric vector. You should use \code{get_variants_index} to retrieve the indexes.}

\item{index_cells}{A numeric vector. Specifies which cells (columns) to retrieve.}

\item{format}{A character. Specifies the return format. Either a \code{list} or a \code{RangedSummarizedExperiment}.}
}
\value{
Default is a named \code{list}. Can return a \code{RangedSummarizedExperiment} if specified.
}
\description{
read_assays_variants retrieves the assays of variants in an h5f.
}
\examples{
#' #Retrieving the NGT matrix for IDH1_R132H and NPM1c
\dontrun{
interesting_variants <- c(IDH1_R132H = "chr2:209113112:C/T", NPM1c = "chr5:170837543:C/CTCTG")
index <- get_variants_index(h5f, interesting_variants)
read_dna_variants(h5f, index_variants = index, included_assays = "NGT")
}
}
