% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_h5}
\alias{read_h5}
\title{Reading a tapestri h5 file}
\usage{
read_h5(path)
}
\arguments{
\item{path}{A character, path to .h5}
}
\value{
h5f
}
\description{
Wrapper function for H5Fopen from rhdf5. Reads in a \code{.h5} file.
}
\examples{
path <- "path/to/.h5"
h5f <- read_h5(path)
}
