% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_protein_counts}
\alias{read_protein_counts}
\title{Reading protein read counts}
\usage{
read_protein_counts(
  h5f,
  transpose = c(TRUE, FALSE),
  index_cells = NULL,
  normalization = c("clr", "raw")
)
}
\arguments{
\item{h5f}{h5f}

\item{transpose}{Boolean. If \code{TRUE}, columns represent proteins and rows represent cells.}

\item{index_cells}{A numeric vector. Specifies which cells to retrieve.}

\item{normalization}{Either \code{clr} or \code{raw}. Defaults to \code{clr}, as implemented by the \code{compositions} package.}
}
\value{
matrix
}
\description{
read_protein_counts retrieves the protein read counts in an h5f.
}
\examples{
path <- "path/to/.h5"
h5f <- read_h5(path)

read_protein_counts(h5f)
}
