% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_scite_output}
\alias{read_scite_output}
\title{Read SCITE output}
\usage{
read_scite_output(file, cell_index_h5f, seed = 123)
}
\arguments{
\item{file}{path to SCITE output file in .gv format.}

\item{cell_index_h5f}{A numeric vector. The index of cells used for scite in the .h5f}

\item{seed}{integer.}
}
\value{
data.table
}
\description{
Designate the clone of each cell based on SCITE. If ambiguous, a random clone from all possible clones is selected.
}
\examples{
\dontrun{
 read_scite_out("scite_output.gv")
}
}
