% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_vep_output}
\alias{read_vep_output}
\title{Read ENSEMBL VEP output}
\usage{
read_vep_output(file)
}
\arguments{
\item{file}{path to vep .txt output file.}
}
\description{
Reads standard VEP output.
}
\examples{
\dontrun{
  write_vep_input(dt.var, "vep_input.txt")
  #run vep locally
  vep <- read_vep_output("path_vep_output.txt")
}
}
