% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_scite_inputs}
\alias{write_scite_inputs}
\title{Write input files for SCITE}
\usage{
write_scite_inputs(h5f, variants, cell_index = NULL, fp_ngt_csv, fp_geneNames)
}
\arguments{
\item{h5f}{h5f}

\item{variants}{A named vector.
You should get the value from the \code{id} column after calling \code{get_variants}.
The names should be meaningful, they are the labels in your inferred clonal architecture .gv file.
Avoid symbols like \code{., -, :} in the names, and use an underscore \code{_} instead.}

\item{cell_index}{A numeric vector. Specifies which cells (columns) to retrieve.}

\item{fp_ngt_csv}{Output file name for NGT matrix .csv}

\item{fp_geneNames}{Output file name for .geneNames}
}
\value{
NULL. Writes two files.
}
\description{
A helper function to write genotype matrix and .geneNames files for SCITE.
}
\references{
https://github.com/cbg-ethz/SCITE

Kuipers J et al. Single-cell sequencing data reveals widespread recurrence and loss of mutational hits in the life histories of tumorus. Genome Research 2017; 27:1885-1894.
}
