% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_vep_input}
\alias{write_vep_input}
\title{Write input files for ENSEMBL VEP}
\usage{
write_vep_input(dt_variants, file)
}
\arguments{
\item{dt_variants}{A (filtered) \code{data.frame} or \code{data.table} of variants derived from \code{get_variants}}

\item{file}{file path to write .txt file.}
}
\value{
NULL. Writes the standard VEP .txt input file.
#' @examples

dt.var <- get_variants(h5f)
write_vep_input(dt.var, "vep_input.txt")
}
\description{
A helper function to write standard input files for ENSEMBL VEP.
}
\references{
https://useast.ensembl.org/info/docs/tools/vep/vep_formats.html
}
